+++ app/build.gradle
@@ -20,6 +20,12 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+    allprojects {
+        repositories {
+            //...
+            maven { url 'https://jitpack.io' }
+        }
+    }
 }
 
 dependencies {
@@ -37,4 +43,5 @@ dependencies {
     implementation 'gun0912.ted:tedpermission:2.2.2'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.google.android.material:material:1.2.0-alpha01'
+    implementation 'com.kroegerama:bottomsheet-imagepicker:1.0.4'
 }
