+++ app/src/main/res/layout/fragment_home_tab.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<GridLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="horizontal"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -11,4 +11,4 @@
             android:layout_height="wrap_content">
         </androidx.recyclerview.widget.RecyclerView>
 
-</GridLayout>
+</RelativeLayout>
