+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,5 @@
     <color name="colorBlack">#000000</color>
     <color name="colorOrange">#FACC2E</color>
     <color name="transparent">#00FFFFFF</color>
+    <color name="colorGray">#F6F6F6</color>
 </resources>
